+++ src/actions.test.js
@@ -6,15 +6,15 @@ import { addRoute, ADD_ROUTE, addRoutes, ADD_ROUTES, delRoute, DEL_ROUTE } from 
 
 describe('addRoute', () => {
   const res = addRoute('foo')
-  const action1 = { payload: { id: 'foo', path: null }, type: ADD_ROUTE }
+  const action1 = { payload: { id: 'foo', pattern: null }, type: ADD_ROUTE }
   const action2 = set('payload.name', 'Foo', action1)
-  const action3 = set('payload.path', '/bar', action1)
+  const action3 = set('payload.pattern', '/bar', action1)
   test('Handle single arg as string', () => {
     expect(res).toEqual(action1)
   })
   test('Handle single arg as object', () => {
     expect(addRoute({ id: 'foo', name: 'Foo' })).toEqual(action2)
-    expect(addRoute({ id: 'foo', path: '/bar' })).toEqual(action3)
+    expect(addRoute({ id: 'foo', pattern: '/bar' })).toEqual(action3)
   })
   test('Handle two arg string, string', () => {
     const res2 = addRoute('foo', '/bar')
@@ -36,10 +36,10 @@ describe('addRoutes', () => {
   const action1 = {
     type: ADD_ROUTES,
     payload: [
-      { id: 'home', path: '/' },
-      { id: 'drawer', path: '/home-drawer' },
-      { id: 'drawerEdit', path: '/home-drawer/:id' },
-      { id: 'image', path: '/image-upload' },
+      { id: 'home', pattern: '/' },
+      { id: 'drawer', pattern: '/home-drawer' },
+      { id: 'drawerEdit', pattern: '/home-drawer/:id' },
+      { id: 'image', pattern: '/image-upload' },
     ],
   }
   test('Object of routes', () => {
@@ -49,10 +49,10 @@ describe('addRoutes', () => {
   const action2 = {
     type: ADD_ROUTES,
     payload: [
-      { id: 'about', path: null },
-      { id: 'contact', path: null },
-      { id: 'favs', path: null },
-      { id: 'showroom', path: null },
+      { id: 'about', pattern: null },
+      { id: 'contact', pattern: null },
+      { id: 'favs', pattern: null },
+      { id: 'showroom', pattern: null },
     ],
   }
   test('Array of routes', () => {